requirements.txt
# TensorFlow 2.10 GPU (最後一個支援 Windows GPU 的版本)
# 需要 CUDA 11.2 和 cuDNN 8.1
tensorflow==2.10.0

# 數值計算庫 (TensorFlow 2.10 需要 numpy < 1.24)
numpy>=1.21.0,<1.24.0

# 資料處理
pandas>=1.3.0,<2.0.0

# 視覺化
matplotlib>=3.4.0,<4.0.0
seaborn>=0.12.0

# 機器學習工具
scikit-learn>=1.0.0,<2.0.0

# 圖像處理
opencv-python>=4.5.0,<5.0.0

# 日期時間處理 (pandas 依賴)
python-dateutil>=2.8.0

# 其他常用依賴
Pillow>=8.0.0
h5py>=3.1.0
protobuf>=3.9.2,<4.0.0

